Add unit test: APIKey(), TextToSpeech()
