+++ Assets/Scripts/Editor/TestScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b0222367a33bcfe41a3635804b4f0ee5
+guid: 6d4022d76d1847f4cbae7e50ac3199b8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
