+++ Assets/Scripts/TextToSpeech/JsonData/AudioConfig.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Assets.Scripts.JsonData.TextToSpeech
+namespace Assets.Scripts.TextToSpeech.JsonData
 {
     [Serializable]
-    class AudioContent
+    class AudioConfig
     {
-        public string audioContent;
+        public string audioEncoding;
     }
 }
