+++ Assets/Scripts/TextToSpeech/JsonData/AudioConfig.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 38ec80bb560dabd4e967531abb4820c3
+guid: 6d390cd0327d0474f8a24a03161c0fa5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
