+++ Assets/Scripts/TextToSpeech/JsonData/TTSRequestBody.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc3a78422260dce4b8aff52a52fa6cbf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
