+++ Assets/Scripts/TextToSpeech/JsonData/Voice.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 697bf75ff4e291349860e321c762461e
+guid: b73f95251146aee4cb6d2dff542d4034
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
